>@< ROBOT C IS FKN COUPLING MEMORY ADDRESSES
Everytime the local variable (made static for visibility) float new_pose_X is modified it modifies probmap[8] (and vice versa it seems). The memory location in probmap shifts around depending on the allocation of other variables (easy way to mess with this is by changing the size of unrelated data queues). Same issue has previously occurred with data in bitmap[] and some long value (ususally in the millions) <- could be the float bits being read as a long.